Fix provider sensitivity in getHostKey.
store.Key returns a key based on the provider the store was opened with. In some cases, particularly when multiple keys are present on the system, this may mismatch store.Cert, which returns a cert from the provider-agnostic cert store.

Instead, lookup the cert we intend to use in the store first, and derive its key directly from it. This should, in theory, imply they will match consistently.

PiperOrigin-RevId: 407359076